Add PortableRoutingPlanFactory
Fixes #148.
